+++ paper/paper.bib
@@ -190,8 +190,7 @@ volume = {141},
 pages = {103779},
 year = {2021},
 issn = {0921-8890},
-doi = {https://doi.org/10.1016/j.robot.2021.103779},
-url = {https://www.sciencedirect.com/science/article/pii/S0921889021000646},
+doi = {10.1016/j.robot.2021.103779},
 author = {Dennis Mronga and Frank Kirchner}
 }
 
