### 2024-04-25 #6530 @ymzayek
@AlexanderFabisch it seems there are also some missing and invalid DOIs per [this comment above](https://github.com/openjournals/joss-reviews/issues/6530#issuecomment-2015637695). Would you be able to update these?
