### 2024-04-25 #6530 @AlexanderFabisch
Hi @ymzayek,

thanks for handling this submission.

> Do you have any suggestion for reviewers?

I looked for expertise in Python, robotics, and machine learning. None of the listed reviewers seems to be an expert in this specific field (which is rather small). So the following reviewers seem to be good candidates (without any preferences):
* mhubii
* gautam-sharma1
* ishaanamahajan
* cst0
* aadi-mishra
* jorgedch
* mrshu
* sea-bass
* giodegas

> @AlexanderFabisch it seems there are also some missing and invalid DOIs per https://github.com/openjournals/joss-reviews/issues/6530#issuecomment-2015637695. Would you be able to update these?

I checked all of the references and they really seem to have no DOIs:

- No DOI given, and none found for title: Active Contextual Policy Search - *JMLR generally does not seem to have DOIs*
- No DOI given, and none found for title: Probabilistic Movement Primitives - *Same for NeurIPS*
- No DOI given, and none found for title: Open3D: A Modern Library for 3D Data Processing - *This was put on arxiv before they started to assign DOIs automatically*
- No DOI given, and none found for title: PyBullet, a Python module for physics simulation f... - *This is not a real publication, but a software*

The invalid DOI should already be fixed with the latest commit on the branch that contains the paper: https://github.com/dfki-ric/movement_primitives/pull/41/commits/89771fb3ce7667a6defe90deb4c7bff3dd41dd34 ([PR](https://github.com/dfki-ric/movement_primitives/pull/41)).
