### 2024-04-29 #6530 @ishaanamahajan
> Hello @mhubii, @gautam-sharma1 and @ishaanamahajan,
>
> Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software)?
>
> JOSS publishes articles about open source research software. The submission I'd like you to review is titled:
>
> [movement_primitives: Imitation Learning of Cartesian Motion with Movement Primitives ](https://github.com/openjournals/joss-reviews/issues/6530)
>
> You can find more information at the top of this Github issue (#6530).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be 23 check items for each single reviewer.
>
> Thank you in advance!

Yes, I am happy to assist. Please let me know how I should move forward.
