### 2024-04-29 #6530 @gautam-sharma1
Hello Yasmin,

Yes, I would be willing to review this article.

Thanks,
Gautam Sharma
_______________________________
Software Engineer
Natick, MA

*+1 480-937-7152 | ***@***.*** ***@***.***>
| www.sharmagautam.com <http://www.sharmagautam.com>*
