### 2024-04-29 #6695 @editorialbot
[REVIEW]: movement_primitives: Imitation Learning of Cartesian Motion with Movement Primitives
**Submitting author:** @AlexanderFabisch (<a href="http://orcid.org/0000-0003-2824-7956">Alexander Fabisch</a>)
**Repository:** https://github.com/dfki-ric/movement_primitives/
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.7.1
