### 2024-05-05 dfki-ric/movement_primitives#42 @AlexanderFabisch
Thanks for reporting. It seems like nosetest is broken in Python 3.11. We will have to switch to pytest.
