+++ .github/workflows/python-package.yml
@@ -8,7 +8,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.7, 3.8]
+        python-version: ["3.7", "3.8", "3.11"]
 
     steps:
       - uses: actions/checkout@v2
@@ -29,6 +29,6 @@ jobs:
           flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
       - name: Test with pytest
         run: |
-          NUMBA_DISABLE_JIT=1 MPLBACKEND=Agg nosetests -sv test/ --with-coverage
+          NUMBA_DISABLE_JIT=1 MPLBACKEND=Agg pytest test/
       - name: Codecov
         uses: codecov/codecov-action@v1.3.2
