+++ README.md
@@ -263,7 +263,7 @@ python -m pip install -e .[test]
 ```
 
 The tests are located in the folder `test/` and can be executed with:
-`python -m nose test`
+`python -m pytest`
 
 This command searches for all files with `test` and executes the functions with `test_*`.
 
