+++ setup.cfg
@@ -6,3 +6,5 @@ cover-html=true
 
 [tool:pytest]
 addopts=--cov-config=.coveragerc --cov=movement_primitives --cov-report html
+testpaths =
+    test
