+++ .github/workflows/python-package.yml
@@ -19,8 +19,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install cython numpy
-          pip install -e .[all] flake8 nose coverage
+          pip install -e .[all,test] flake8
       - name: Lint with flake8
         run: |
           # stop the build if there are Python syntax errors or undefined names
