### 2024-05-05 417d488
Remove nose dependency
