+++ test/test_canonical_system.py
@@ -1,7 +1,7 @@
 import numpy as np
 from movement_primitives.dmp._canonical_system import canonical_system_alpha
 from movement_primitives.dmp._canonical_system import phase as phase_python
-from nose.tools import assert_almost_equal
+import pytest
 
 
 def test_phase_cython():
@@ -13,4 +13,4 @@ def test_phase_cython():
     for t in np.linspace(0, 1, 101):
         z_python = phase_python(t, alpha, goal_t, start_t, int_dt)
         z_cython = phase_cython(t, alpha, goal_t, start_t, int_dt)
-        assert_almost_equal(z_cython, z_python)
+        assert z_cython == pytest.approx(z_python)
