### 2024-05-05 dfki-ric/movement_primitives#42 @AlexanderFabisch
@gautam-sharma1 I opened a PR that replaces nose by pytest: #43 . Could you confirm that it works? Then I will merge it to main.
