### 2024-05-05 #6695 @gautam-sharma1
Created an issue related to test (https://github.com/dfki-ric/movement_primitives/issues/42), author resolved it, so marking tests as complete.
