### 2024-05-05 dfki-ric/movement_primitives#44 @gautam-sharma1
JOSS suggestion: Update Copyright year in LICENSE
It would be good to update the copyright year to the current year in the LICENSE file
