### 2024-05-05 dfki-ric/movement_primitives#45 @gautam-sharma1
JOSS suggestion: Statement of need
After reading the draft, I got a sense that movement_primitives does claim to be faster than other repos in the same space. If that's the case, then it will be good to add that to the "Statement of Need".

Also, is there any way I can reproduce the results?
