### 2024-05-05 dfki-ric/movement_primitives#46 @gautam-sharma1
JOSS suggestion: Summary
It will be good to add a figure describing what exactly the movement primitives are trying to learn or imitate for non specialist audience. It can be a graph similar to the one in the README. Something to let the reader visually know what the goal policy is that the movement primitives are trying to imitate.
