### 2024-05-05 dfki-ric/movement_primitives#47 @gautam-sharma1
JOSS suggestion: Add statement of need to documentation
Can you please add a statement of need in the documentation as well?
