### 2024-05-05 #6695 @gautam-sharma1
Created few issues in the main repo:
1. https://github.com/dfki-ric/movement_primitives/issues/44
2. https://github.com/dfki-ric/movement_primitives/issues/45
3. https://github.com/dfki-ric/movement_primitives/issues/46
4. https://github.com/dfki-ric/movement_primitives/issues/47
