### 2024-05-05 dfki-ric/movement_primitives#44 @AlexanderFabisch
I removed the years from the copyright as it does not seem to be necessary (https://github.com/dfki-ric/movement_primitives/commit/d975ae03279889e1b375985db442b767d01df779)
