### 2024-05-05 dfki-ric/movement_primitives#45 @AlexanderFabisch
> After reading the draft, I got a sense that movement_primitives does claim to be faster than other repos in the same space. If that's the case, then it will be good to add that to the "Statement of Need".

It is fast for Python, but dmpbbo in C++ is faster for most cases with few dimensions. In high-dimensional cases this library seems to be better than dmpbbo though. So I don't know if I can claim this as an advantage of this library over others. In the summary I mention that Cython is used to speed up online execution. What do you think?

> Also, is there any way I can reproduce the results?

I just uploaded most of the benchmark code to this gist: https://gist.github.com/AlexanderFabisch/a69534fe425d305d7550eb526825cf2f

The C++ code for dmpbbo is currently missing though. I'll need some time to recover it.
