### 2024-05-05 dfki-ric/movement_primitives#47 @AlexanderFabisch
I added it in https://github.com/dfki-ric/movement_primitives/commit/68e2b4623a71cdc4c54f8af17899eecebfd2cc4e
