+++ README.md
@@ -20,6 +20,8 @@ addition, the library provides tools for data analysis and movement evaluation.
 It can be installed directly from
 [PyPI](https://pypi.org/project/movement-primitives/).
 
+<img src="doc/source/_static/summary.svg" width="100%" />
+
 ## Content
 
 * [Statement of Need](#statement-of-need)
