### 2024-05-08 dfki-ric/movement_primitives#46 @AlexanderFabisch
That's a good idea. I made a figure illustrating the imitation learning process to the readme and the paper. I will close the issue, but feel free to open it if you have additional comments.
