### 2024-05-18 #6695 @ishaanamahajan
Overall, the contribution has a clear impact, and a meaningful contribution to the robotics, and to the open-source software community. I also found the writing to be pretty strong. However there are a couple of things I feel should be brushed upon/need improvements before publication -

1. The statement of need lacks depth in my opinion. For example, a comparison with neural networks is brought up, however concrete examples illustrating the differences are missing. Another suggestion (a modification of the former suggestion) would be to provide the current landscape in the field, and give specific examples on why movement primitives are better (though they are given briefly in the 'Design and Features' section, I feel the statement of need would also benefit with points/arguments along these lines).

2. In the Installation instructions, I feel they could be separated better. For example, 'Installation with pip' could be a section, 'Installation with Git' could another. As a user of this library, I would only want to read the installation section that I find is relevant to me, rather than reading everything and figuring out which commands I need to run, and which ones I do not.

Apart from that, I found the library useful, and the examples/demos given easy to run and interpret.
