### 2024-05-18 aa476b9
Update installation instructions
