### 2024-05-18 #6695 @AlexanderFabisch
Thank you for the review @ishaanamahajan .

> The statement of need lacks depth in my opinion. For example, a comparison with neural networks is brought up, however concrete examples illustrating the differences are missing. Another suggestion (a modification of the former suggestion) would be to provide the current landscape in the field, and give specific examples on why movement primitives are better (though they are given briefly in the 'Design and Features' section, I feel the statement of need would also benefit with points/arguments along these lines).

I updated the statement of need in the paper and in the readme: https://github.com/dfki-ric/movement_primitives?tab=readme-ov-file#statement-of-need

Please let me know if it could be improved.

> In the Installation instructions, I feel they could be separated better. For example, 'Installation with pip' could be a section, 'Installation with Git' could another. As a user of this library, I would only want to read the installation section that I find is relevant to me, rather than reading everything and figuring out which commands I need to run, and which ones I do not.

I updated the section in the readme: https://github.com/dfki-ric/movement_primitives?tab=readme-ov-file#install-library

The readme is also included in the Sphinx documentation. So with the next release this should also be updated.
