### 2024-05-26 #6695 @ymzayek
@AlexanderFabisch can you have a look at the additional author tasks in https://github.com/openjournals/joss-reviews/issues/6695#issuecomment-2132287228 ?
