+++ _sources/_apidoc/movement_primitives.dmp.CouplingTermDualCartesianTrajectory.rst.txt
@@ -7,22 +7,3 @@
    :members:
    :show-inheritance:
    :inherited-members:
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~CouplingTermDualCartesianTrajectory.__init__
-      ~CouplingTermDualCartesianTrajectory.couple_distance
-      ~CouplingTermDualCartesianTrajectory.coupling
-      ~CouplingTermDualCartesianTrajectory.imitate
-   
-   
-
-   
-   
-   
