+++ _sources/_apidoc/movement_primitives.kinematics.Kinematics.rst.txt
@@ -7,20 +7,3 @@
    :members:
    :show-inheritance:
    :inherited-members:
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~Kinematics.__init__
-      ~Kinematics.create_chain
-   
-   
-
-   
-   
-   
