+++ _sources/_apidoc/movement_primitives.base.PointToPointMovement.rst.txt
@@ -7,22 +7,3 @@
    :members:
    :show-inheritance:
    :inherited-members:
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~PointToPointMovement.__init__
-      ~PointToPointMovement.configure
-      ~PointToPointMovement.n_steps_open_loop
-      ~PointToPointMovement.step
-   
-   
-
-   
-   
-   
