+++ _sources/_apidoc/movement_primitives.dmp.CouplingTermObstacleAvoidance2D.rst.txt
@@ -7,20 +7,3 @@
    :members:
    :show-inheritance:
    :inherited-members:
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~CouplingTermObstacleAvoidance2D.__init__
-      ~CouplingTermObstacleAvoidance2D.coupling
-   
-   
-
-   
-   
-   
