+++ _sources/_apidoc/movement_primitives.dmp.CouplingTermPos1DToPos1D.rst.txt
@@ -7,20 +7,3 @@
    :members:
    :show-inheritance:
    :inherited-members:
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~CouplingTermPos1DToPos1D.__init__
-      ~CouplingTermPos1DToPos1D.coupling
-   
-   
-
-   
-   
-   
