+++ _sources/api.rst.txt
@@ -25,25 +25,24 @@ You can also search for specific modules, classes, or functions in the
    ~movement_primitives.dmp.DMPWithFinalVelocity
    ~movement_primitives.dmp.CartesianDMP
    ~movement_primitives.dmp.DualCartesianDMP
-   ~movement_primitives.dmp.DMPBase
-   ~movement_primitives.dmp.WeightParametersMixin
+
+Coupling Terms
+--------------
+
+Coupling terms can be used to adapt the movement generated by a DMP online.
+Not all coupling terms are compatible with all types of DMPs.
+
+.. autosummary::
+   :toctree: _apidoc/
+   :template: class.rst
+
    ~movement_primitives.dmp.CouplingTermPos1DToPos1D
    ~movement_primitives.dmp.CouplingTermObstacleAvoidance2D
    ~movement_primitives.dmp.CouplingTermObstacleAvoidance3D
    ~movement_primitives.dmp.CouplingTermPos3DToPos3D
    ~movement_primitives.dmp.CouplingTermDualCartesianPose
-   ~movement_primitives.dmp.CouplingTermObstacleAvoidance3D
    ~movement_primitives.dmp.CouplingTermDualCartesianDistance
    ~movement_primitives.dmp.CouplingTermDualCartesianTrajectory
-
-.. autosummary::
-   :toctree: _apidoc/
-
-   ~movement_primitives.dmp.dmp_transformation_system
-   ~movement_primitives.dmp.canonical_system_alpha
-   ~movement_primitives.dmp.phase
-   ~movement_primitives.dmp.obstacle_avoidance_acceleration_2d
-   ~movement_primitives.dmp.obstacle_avoidance_acceleration_3d
 
 
 :mod:`movement_primitives.promp`
