+++ _sources/_apidoc/movement_primitives.spring_damper.SpringDamperOrientation.rst.txt
@@ -7,24 +7,3 @@
    :members:
    :show-inheritance:
    :inherited-members:
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~SpringDamperOrientation.__init__
-      ~SpringDamperOrientation.configure
-      ~SpringDamperOrientation.n_steps_open_loop
-      ~SpringDamperOrientation.open_loop
-      ~SpringDamperOrientation.reset
-      ~SpringDamperOrientation.step
-   
-   
-
-   
-   
-   
