+++ _sources/_apidoc/movement_primitives.kinematics.Chain.rst.txt
@@ -7,26 +7,3 @@
    :members:
    :show-inheritance:
    :inherited-members:
-
-   
-   .. automethod:: __init__
-
-   
-   .. rubric:: Methods
-
-   .. autosummary::
-   
-      ~Chain.__init__
-      ~Chain.ee_pose_error
-      ~Chain.forward
-      ~Chain.forward_trajectory
-      ~Chain.inverse
-      ~Chain.inverse_trajectory
-      ~Chain.inverse_with_random_restarts
-      ~Chain.local_inverse_with_random_restarts
-   
-   
-
-   
-   
-   

[…]
