### 2024-05-29 #6695 @AlexanderFabisch
The title of the Zenodo archive was automatically generated by the Github-Zenodo interface. I'd prefer to keep both, the name of the Zenodo archive and the name of the paper. Would that not be possible?
