### 2024-05-30 #6695 @ymzayek
> The title of the Zenodo archive was automatically generated by the Github-Zenodo interface. I'd prefer to keep both, the name of the Zenodo archive and the name of the paper. Would that not be possible?

That is not up to me but as I understand it it should match
@crvernon maybe can chime in on that?
