### 2024-05-30 #6695 @crvernon
:wave: @AlexanderFabisch and @ymzayek

> > The title of the Zenodo archive was automatically generated by the Github-Zenodo interface. I'd prefer to keep both, the name of the Zenodo archive and the name of the paper. Would that not be possible?
>
> That is not up to me but as I understand it it should match @crvernon maybe can chime in on that?

The title of the paper and the title of the Zenodo archive do need to match exactly.  The Zenodo title can be edited to accommodate this need.  Or you can supply a CITATION.cff file in the root of your GitHub repository that controls what Zenodo uses for metadata; which is a good approach.  Along with the title matching, the Zenodo archive should also 1) list the same author list in the same order as the paper and 2) declare the same license that is in your source GitHub repository.

You can check out some recent JOSS publications to see how authors have formed their CITATION.cff file.

Thanks!
