### 2024-05-31 #6695 @crvernon
@AlexanderFabisch the paper is looking great!

I only need one correction before we publish:

- [ ] LINE 97:  The citation in the statement "size through a coupling term defined by (Gams et al., 2013)." should appear as "size through a coupling term defined by Gams et al.(2013)."

No reason to do another release.  Just correct this in your paper and let me know when it is done.

Thanks!
