+++ paper/paper.md
@@ -158,7 +158,7 @@ from PyPI.
 rotating an object with two hands. The movement is generated by a dual
 Cartesian DMP trained on a demonstrated rotation movement. The width of the
 object is known. Hence, it can easily be adapted for similar objects with a
-different size through a coupling term defined by [@Gams2013].
+different size through a coupling term defined by @Gams2013.
 
 ![RH5 Manus [@Boukheddimi2022] rotating a compact solar panel.\label{fig:rh5}](rh5_panel_0){ width=50% }
 
